feat(rps): Make app testable
* build: add jest for testing

I've used Jest before so let's just use it and keep the learning curve
short, although a more minimal framework may be simpler.

* build: install babel to support ES6 modules in tests

I really don't want to bloat things up with Babel but it's necessary for
using jest; the other option is to use "require" but module.exports will
not work within the browser. It appears that raw jest depends on Babel
anyways.

* refactor: use ES6 modules for testing
* refactor: rename some fns to make tests more readable
* refactor: expose state globally to window for debugging
* feat: add basic tests as a PoC for testability